+++ netbox/circuits/forms/bulk_import.py
@@ -4,6 +4,7 @@ from django.utils.translation import gettext_lazy as _
 from circuits.choices import *
 from circuits.models import *
 from dcim.models import Site
+from netbox.choices import DistanceUnitChoices
 from netbox.forms import NetBoxModelImportForm
 from tenancy.models import Tenant
 from utilities.forms.fields import CSVChoiceField, CSVModelChoiceField, SlugField
@@ -94,6 +95,12 @@ class CircuitImportForm(NetBoxModelImportForm):
         choices=CircuitStatusChoices,
         help_text=_('Operational status')
     )
+    distance_unit = CSVChoiceField(
+        label=_('Distance unit'),
+        choices=DistanceUnitChoices,
+        required=False,
+        help_text=_('Distance unit')
+    )
     tenant = CSVModelChoiceField(
         label=_('Tenant'),
         queryset=Tenant.objects.all(),
@@ -106,7 +113,7 @@ class CircuitImportForm(NetBoxModelImportForm):
         model = Circuit
         fields = [
             'cid', 'provider', 'provider_account', 'type', 'status', 'tenant', 'install_date', 'termination_date',
-            'commit_rate', 'description', 'comments', 'tags'
+            'commit_rate', 'distance', 'distance_unit', 'description', 'comments', 'tags'
         ]
