+++ netbox/dcim/models/racks.py
@@ -16,13 +16,13 @@ from dcim.constants import *
 from dcim.svg import RackElevationSVG
 from netbox.choices import ColorChoices
 from netbox.models import OrganizationalModel, PrimaryModel
+from netbox.models.mixins import WeightMixin
 from netbox.models.features import ContactsMixin, ImageAttachmentsMixin
 from utilities.conversion import to_grams
 from utilities.data import array_to_string, drange
 from utilities.fields import ColorField, NaturalOrderingField
 from .device_components import PowerPort
 from .devices import Device, Module
-from .mixins import WeightMixin
 from .power import PowerFeed
 
 __all__ = (
