+++ netbox/netbox/tables/columns.py
@@ -35,6 +35,7 @@ __all__ = (
     'ContentTypesColumn',
     'CustomFieldColumn',
     'CustomLinkColumn',
+    'DistanceColumn',
     'DurationColumn',
     'LinkedCountColumn',
     'MarkdownColumn',
@@ -691,3 +692,16 @@ class ChoicesColumn(tables.Column):
             value.append(f'({omitted_count} more)')
 
         return ', '.join(value)
+
+
+class DistanceColumn(TemplateColumn):
+    """
+    Distance with template code for formatting
+    """
+    template_code = """
+    {% load helpers %}
+    {% if record.distance %}{{ record.distance|floatformat:"-2" }} {{ record.distance_unit }}{% endif %}
+    """
+
+    def __init__(self, template_code=template_code, order_by='_abs_distance', **kwargs):
+        super().__init__(template_code=template_code, order_by=order_by, **kwargs)
