--- netbox/wireless/tables/template_code.py
@@ -1,4 +1,0 @@
-WIRELESS_LINK_DISTANCE = """
-{% load helpers %}
-{% if record.distance %}{{ record.distance|floatformat:"-2" }} {{ record.distance_unit }}{% endif %}
-"""
