+++ netbox/wireless/tables/wirelesslink.py
@@ -4,7 +4,6 @@ import django_tables2 as tables
 from netbox.tables import NetBoxTable, columns
 from tenancy.tables import TenancyColumnsMixin
 from wireless.models import *
-from .template_code import WIRELESS_LINK_DISTANCE
 
 __all__ = (
     'WirelessLinkTable',
@@ -37,10 +36,7 @@ class WirelessLinkTable(TenancyColumnsMixin, NetBoxTable):
         verbose_name=_('Interface B'),
         linkify=True
     )
-    distance = columns.TemplateColumn(
-        template_code=WIRELESS_LINK_DISTANCE,
-        order_by=('_abs_distance')
-    )
+    distance = columns.DistanceColumn()
     tags = columns.TagColumn(
         url_name='wireless:wirelesslink_list'
     )
